core: introduce new color depth factor.
The original image creation function assumed that the input image has
a 8bit color depth.

Whenever an image with 16bit color depth was used as input, the output
image contained "random noise".

Bugfix: calculate a color depth factor that is used in the pixel
assignment logic for the output image buffer.