Fix jpeg cropping edge case
TurboJpeg expects cropping regions to be multiples of the MCU size. In
some cases it was possible that the cropping region was larger than the
image due to rounding artifacts (less than one MCU). This MR checks if
the resulting region would exceed the image and makes it one MCU width
and/or height smaller if needed.