Fix crash during JPEG2000 encoding.
A misunderstanding in the way JNR-FFI worked led to a hard-to-debug
segmentation fault crash when encoding JPEG2000 images concurrently.

The bug was in our call to `opj_image_create` during encoding. This
function takes a pointer to an array of `opj_image_comptparm` structs.
On the Java side, we allocated those structs through JNR-FFI's `Runtime`
and passed the address of the first member of the array to the function.
This was done under the assumption that the memory allocated by JNR-FFI
for the array members would be the same memory that gets passed over the
FFI, and, since we confirmed that JNR-FFI allocates those array members
as a single contiguous memory segment, it would be safe to do so.
However, this is based on a significant misunderstanding how passing
memory between the JVM and the native code works with JNR-FFI. It seems
that when a pointer is passed over the FFI, a copy is created for
the memory it is pointing at, and when the function returns, that
temporary memory is copied back to the JVM memory. It seems that only
the memory that is backed by a `Pointer` instance on the JVM side is
safe from deallocation during the function call. The solution to the
crashes is thus to simply create a pointer for every struct member to
make sure it actually gets copied over and is not deallocated at some
point. Open questions remain: Why does the segfault only happen
occisaionaly (high concurrency + large images on my dev machine). And
why doesn't it happen when we attach a debugger?

This fix also points at a potential avenue for performance improvements:
How much unneeded copying is going on that we could skip via `@In`,`@Out` or
`@Direct` annotations?