openjpeg: Work around concurrency issue in jnr.ffi.Struct#arrayOf
This is another attempt at fixing the bug mentioned in an earlier
commit. The previous attempt was already headed in the right direction
but didn't actually work.
This next attempt seems to be more promising, at least I couldn't
trigger it when hammering my encoding test on 8 threads for ~3h, that
would pretty reliably crash previously.

The gory details:
`Struct.arrayOf` allocates a continuous chunk to back an array of
structures. JNR's `TransientMemoryAllocator` does not allocate memory
directly for small allocations (<=256 bytes), but uses a custom
page-based allocation strategy. When allocation for a new page fails,
it triggers a full GC cycle and clears up all 'old and unused' pages.
However, there seems to be a bug involving concurrency and garbage
collection, that leads to this cleanup process freeing memory that is
still in use.
To work around this, we force arrays of structs to be always backed by
a memory block of at least 512 bytes, thereby bypassing the transient
memory allocator and using JNR's direct memory allocation strategy
that follows different rules for freeing unused memory.