feat(imageio-openjpeg): enable raster read on `OpenJp2ImageReader`
The class now overrides the raster functions from `ImageReader` class.
The `canReadRaster` returns true enabling the caller to call `readRaster` and `readTileRaster`.

The readRaster and readTileRaster call the normal read functions and then return the raster with `getData` call on the buffered image.

Test cases are added to compare the rasters on expected and actual images with the option of tolerance in the diff of samples if the quality of images is low.